requirement.txt
asgiref==3.5.2
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.
ipython==7.24.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mysqlclient==2.0.3
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
ptyprocess==0.7.0
Pygments==2.12.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytz==2022.1
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
traitlets==5.3.0
typing_extensions==4.2.
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5